# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Grab Restaurant API",
    "description": "All RESTful API Endpoint for Grab Restaurant",
    "version": "0.0.1"
  },
  "paths": {
    "/api/auth/signUp": {
      "post": {
        "tags": ["Authentication"],
        "description": "Sign in to Grab Restaurant",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "number" },
                    "username": { "type": "string" },
                    "email": { "type": "string" },
                    "accessToken": { "type": "string" }
                  }
                },
                "example": {
                  "id": 4,
                  "username": "admin",
                  "email": "admin@gmail.com",
                  "accessToken": "eyJhbGc....S_2b42-zD4"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signIn": {
      "post": {
        "tags": ["Authentication"],
        "description": "Sign in to Grab Restaurant",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully signed in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "id": 3,
                    "username": "username",
                    "email": "email@webmail.npru.ac.th",
                    "roles": ["ADMIN"],
                    "accessToken": "eyJhbGc....S_2b42-zD4"
                  }
                }
              }
            }
          }
        }
      }
    },

    "/Carcenters": {
      "get": {
        "tags": ["Carcenters"],
        "responses": {
          "200": {
            "description": "Successfully restaurant data"
          }
        }
      },
      "post": {
        "tags": ["Carcenters"],
        "description": "Add a Car to Carcenters",
        "headers": {
          "Authorization": "Bearer YourAccessTokenHere"
        },
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "An authorization header"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "brand": { "type": "string" },
                "model": { "type": "string" },
                "price": { "type": "string" },
                "imageFirst": { "type": "string" },
                "imageSecond": { "type": "string" },
                "imageThird": { "type": "string" },
                "firstprimarycolor": { "type": "string" },
                "secondprimarycolor": { "type": "string" },
                "thirdprimarycolor": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully add a Car to Carcenters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "car": {
                      "type": "object",
                      "properties": {
                        "brand": { "type": "string" },
                        "model": { "type": "string" },
                        "price": { "type": "string" },
                        "imageFirst": { "type": "string" },
                        "imageSecond": { "type": "string" },
                        "imageThird": { "type": "string" },
                        "firstprimarycolor": { "type": "string" },
                        "secondprimarycolor": { "type": "string" },
                        "thirdprimarycolor": { "type": "string" }
                      }
                    }
                  }
                },
                "example": {
                  "message": "Car added successfully",
                  "car": {
                    "brand": "Ferrari",
                    "model": "488 Pista Spider",
                    "price": "11,600,000",
                    "imageFirst": "https://images.autofun.co.th/file1/16d84793ae9a41aa9d8baddef4266881_912x516.jpg",
                    "imageSecond": "https://images.autofun.co.th/file1/16d84793ae9a41aa9d8baddef4266881_912x516.jpg",
                    "imageThird": "https://images.autofun.co.th/file1/16d84793ae9a41aa9d8baddef4266881_912x516.jpg",
                    "firstprimarycolor": "#ffffff",
                    "secondprimarycolor": "#000099",
                    "thirdprimarycolor": "#000000"
                  }
                }
              }
            }
          }
        }
      }
    },
    
    "/Carcenters/{id}": {
      "get": {
        "tags": ["Carcenters"],
        "description": "Get Carcenters by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },


          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "An authorization header"
          }

        ],
        "responses": {
          "200": {
            "description": "Successfully get Carcenters by Id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "Img": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Carcenters"],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "An authorization header"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "brand": { "type": "string" },
                "model": { "type": "string" },
                "price": { "type": "string" },
                "imageFirst": { "type": "string" },
                "imageSecond": { "type": "string" },
                "imageThird": { "type": "string" },
                "firstprimarycolor": { "type": "string" },
                "secondprimarycolor": { "type": "string" },
                "thirdprimarycolor": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated Carcenters"
          }
        }
      },
      "delete": {
        "tags": ["Carcenters"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "An authorization header"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted Carcenters"
          }
        }
      }
    }
  }
}